requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
numpy==2.1.1
openai==1.45.1
packaging==24.1
pandas==2.2.2
plotly==5.24.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.9.11
requests==2.32.3
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
